requirements_dev.txt
coverage
einops
flake8
numpy
protobuf
pytest
pytorch-lightning
requests
torchmetrics
urllib3
xbatcher
xarray[io]
dask[complete]
